canonical ordering for + and * construction.
This weakly orders the arguments to + and * by type. Across type boundaries, we are now expecting these to be commutative. This may be stronger medicine than we want to drink; if it were just a matter of corralling numbers, it might be better to stick to that, since that's less likely to produce troube in the future.