tiny steps toward adding rule simplification
The most important thing we figured out here was how to let a rule consequence term be a function of the frame instead of just a lookup of a symbol in the frame; together with a shortcut so that the old conseqence syntax could work. With this we managed to keep "eval" at bay.