made poly library a little more rigid about type.
It was bothering me how often we checked for base. So now we are less concerned about lowering constant polynomials back to the base ring, especially since we don't know what the base ring is. This feels a little better from the point of view of type. There is some chance this should be pushed further.