Trial balloon change of two operations to defmulti
instead of Scmutils defhandler. Promising so far. One little-used
unary operation (tan) and one heavily used binary operation (+)
were selected for this experiment.

We discovered that operand order-canonicalization is helping
to control the number of defmethod specializations we need.
This is interesting and warrants further study. We don't want
the dispatch hierarchy to become a tangle, but we wonder whether
operand order sorting is a good thing in the long run. The trade-
off may in fact be correctly set right now.