some overdue work on polynomials.
The main thing is that when riaising an expression to polynomial form, we now raise all the constants to constant polynomials. This allows us to assume that all parameters to polynomial arithmetic functions will themselves be polynomials, (even in the one case where that is less convenient: expt).

The benefit is we can burn down all the special casing of non-polynomial arguments to poly arithmetic functions: it's cleaner