i'm making a note here:
huge (?) success.

This begins to weaken the "RF-only" notion of arithmetic in the RF simplifier engine. E.g., we no longer lift constants to polynomials and don't lift polynomials to rational functions with denominator one.

The abstractions feel better handled. But the gcd is still too expensive to simplify the rigid-rotation example, which is disappointing.

some more diagnostics are needed to make sure we're not doing gcds unless we absolutely have to.