improve handling of :arity message
It used to be that v/argument-kind would map keywords to themselves, acting as if the set of keywords were a sea of self-describing unit types. This worked, but felt like bending the rules. Now, argument type of a :keyword is Keyword, and we set up a method for all keywords (which may react to the actual keyword seen).