refactor:generic operations should all be multifns
There are 3 kinds of polymorphism used currently in this library. (1) the Value protocol, (2) the generic multifns, and (3) a disturbingly large number of ‘if’ expressions in functions implementing generic operations. This is out of hand. Instead, we are going to centralize on multifns, moving away from the value protocol. We have established a new lattice of types, and simplified the generation of multifn specializations, so that it becomes (or will become) easier to reason about how values of varous types are combined by generic operations.

zero-like is the first Value protocol method to fall.

A great reorganization of the way the expression type is handled has begun here.