Having keywords type to themselves was a mistake
This caused some infinite type-loops and made printing keyword output problematic. It had some “surprising” effects, and not the good kind of surprise, so we have to backtrack on that.

Instead, we type them by their class (as before), and deal more carefully with the :arity and :name special messages for functions.