Speedup for the pattern matching simplifier
Macros are hard. Due to bad macrology on my part I had left the
compilation of individual rule patterns within the body of generated
rulesets, so that the rule patterns were compiled each time the
ruleset was applied, defeating the purpose of the compilation. This
moves the pattern compilation into a let-block outside the ruleset
function definition and moves to a tail-recursive iteration over
included rules instead of a huge if-chain. It seems much nicer this
way.