Walk back moving g/zero? from protocol to MultiFn.
It turns out that protocol implementations are much faster to invoke
than MultiFn methods. Protocol dispatch can use Java native
polymorphism. Multifns require querying a database of implementations
while a lock is held: this is noticeably more expensive, and in the
GCD calculations it hurts. Therefore we move polynomial zero-testing
back to a function, as a first step.